fix(mobile): migrate to kotlinx-datetime 0.7.1 + kotlin.time.Clock
kotlinx.datetime.Clock was removed; switch all usages to kotlin.time.Clock.
kotlinx.datetime.Instant deprecated; switch to kotlin.time.Instant.
LocalDate.toEpochDays() now returns Long (was Int) — adjust call sites.
LocalDateTime.dayOfMonth/monthNumber deprecated; use day/month with
.ordinal for the 0-based index into Polish-month arrays.

+++ mobile/app-ui/src/commonMain/kotlin/com/poziomki/app/ui/feature/home/EventsScreen.kt
@@ -568,7 +568,7 @@ private fun WeekEventsContent(
                 .groupBy { eventDateKey(it.startsAt) }
                 .toSortedMap()
         }
-    val collapsedDays = remember { mutableStateMapOf<Int, Boolean>() }
+    val collapsedDays = remember { mutableStateMapOf<Long, Boolean>() }
 
     LazyColumn(
         modifier =
